Revert "subsection tree and modularization"
