+++ .gitignore
@@ -3,5 +3,3 @@ dist/prod/*.js
 dist/prod/**
 node_modules
 builds/**
-.ideas
-.DS_STORE
