--- .idea/jsLibraryMappings.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="JavaScriptLibraryMappings">
-    <file url="file://$PROJECT_DIR$" libraries="{client-app/node_modules}" />
-    <file url="PROJECT" libraries="{client-app/node_modules}" />
-    <includedPredefinedLibrary name="WebGL" />
-  </component>
-</project>
