--- src/js/modules/articles/actionTypes.js
@@ -1,3 +1,0 @@
-export const FETCH_ARTICLE_PENDING = 'articles/FETCH_ARTICLE_PENDING';
-export const FETCH_ARTICLE_REJECTED = 'articles/FETCH_ARTICLE_REJECTED';
-export const FETCH_ARTICLE_FULFILLED = 'articles/FETCH_ARTICLE_FULFILLED';
