--- src/js/modules/articles/components/index.js
@@ -1,1 +1,0 @@
-export { default as ArticlePage } from "./ArticlePage";
