--- src/js/modules/articles/constants.js
@@ -1,4 +1,0 @@
-export default
-{
-    NAME :  'articles',
-};
