--- src/js/modules/articles/index.js
@@ -1,7 +1,0 @@
-import * as actions from './actions';
-import * as components from './components';
-import * as constants from './constants';
-import reducer from './reducer';
-import * as selectors from './selectors';
-
-export default { actions, components, constants, reducer, selectors };
