+++ src/js/modules/core/actionTypes.js
@@ -1,3 +1,1 @@
-export const SET_LANGUAGE = 'core/SET_LANGUAGE';
-export const REFRESH_WINDOW_DIMENSIONS = 'core/REFRESH_WINDOW_DIMENSIONS';
-
+export const FETCH_ARTICLE = 'core/FETCH_ARTICLE'
