--- src/js/modules/core/components/index.js
@@ -1,1 +1,0 @@
-export { default as HomePage } from "./HomePage";
