+++ src/js/modules/core/components/style/LanguageMenuStyle.js
@@ -1,0 +1,29 @@
+const styleSheet =
+{
+    languageContainer :
+    {
+        marginLeft    : '12pt',
+        display       : 'flex',
+        alignItems    : 'center',
+        flexDirection : 'row',
+        color         : '#FFFFFF'
+    },
+    languageButtonIcon :
+    {
+        width    : '32px',
+        position : 'relative',
+        color    : '#FFFFFF'
+    },
+    languageButtonContainer :
+    {
+        width      : '32px',
+        marginLeft : '16px'
+    },
+    languageTextSpan :
+    {
+        fontWeight : 900,
+        marginRight: '-16px'
+    }
+};
+
+export default styleSheet
