+++ src/js/modules/core/constants.js
@@ -1,4 +1,4 @@
 export default
 {
-    NAME :  'core',
-};
+    NAME :  'core'
+}
