--- src/js/modules/sections/actionTypes.js
@@ -1,3 +1,0 @@
-export const FETCH_SECTION_PENDING = 'sections/FETCH_SECTION_PENDING';
-export const FETCH_SECTION_REJECTED = 'sections/FETCH_SECTION_REJECTED';
-export const FETCH_SECTION_FULFILLED = 'sections/FETCH_SECTION_FULFILLED';
