--- src/js/modules/sections/components/index.js
@@ -1,1 +1,0 @@
-export { default as SectionPage } from './SectionPage';
