--- src/js/modules/sections/constants.js
@@ -1,4 +1,0 @@
-export default
-{
-    NAME :  'sections',
-}
