+++ src/js/reducers.js
@@ -1,13 +1,9 @@
-import combineReducers from 'redux/lib/combineReducers'
-import { routerReducer } from 'react-router-redux/reducer'
-import core from './modules/core'
-import articles from './modules/articles'
-import sections from './modules/sections'
+import combineReducers     from 'redux/lib/combineReducers'
+import { routerReducer }   from 'react-router-redux/reducer'
+import core                from './modules/core'
 
 export default combineReducers(
-  {
-    [core.constants.NAME]: core.reducer,
-    [articles.constants.NAME]: articles.reducer,
-    [sections.constants.NAME]: sections.reducer,
-    router: routerReducer
-  });
+{
+    [core.constants.NAME]              : core.reducer,
+    router                             : routerReducer
+});
