+++ src/js/versionInfo.js
@@ -3,5 +3,5 @@
  *           without modifying formatting for automation scripts
  * @type {{VERSION: string}}
  */
-let VersionInfo = { VERSION : '0.0.715' };
+let VersionInfo = { VERSION : '0.0.498' };
 export default VersionInfo
